Se mejoro la función que hacer que puedas salir de la app, y que no genere choques con las otras pantallas
