+++ app/(recepcionist)/_layout.tsx
@@ -1,6 +1,11 @@
 import RecepcionistNavBar from '@/components/recepcionista/RecepcionistNavBar';
+import { BackHandlerProvider } from '@/contexts/BackHandlerContext';
 import React from 'react';
 
 export default function RecepcionistLayout() {
-  return <RecepcionistNavBar />;
+  return (
+    <BackHandlerProvider>
+      <RecepcionistNavBar />
+    </BackHandlerProvider>
+  );
 }
